requirements.txt
openai
pinecone[gRPC]
python-dotenv
streamlit